refactor(test): extract async replication fixtures to conftest
Move fast_forward helper and model fixtures (first_model, second_model)
from test_async_replication.py to conftest.py to enable reuse across
multiple async replication test files
